fix-search

+++ src/views/pages/users/User.js
@@ -8,17 +8,49 @@ const User = () => {
   const auth = useSelector((state) => state.auth)
   const [users, setUsers] = useState([])
   const [pageIndex, setPageIndex] = useState(1)
-  const [pageSize] = useState(5)
+  const [pageSize] = useState(10)
   const [visible, setVisible] = useState(false)
   const [search, setSearch] = useState('')
 
-  const totalPage = Math.ceil(users.length / pageSize)
+  const filteredUsers = users.filter((user) => user.username.includes(search))
+  const totalPage = Math.ceil(filteredUsers.length / pageSize)
+
+  // const handleDelete = async (id) => {
+  //   const confirmed = window.confirm('Are you sure you want to delete this user?')
+  //   if (confirmed) {
+  //     deleteUser(id).then((data) => {
+  //       setUsers(() => users.filter((user) => user._id !== id))
+  //       alert('User deleted successfully.')
+  //     })
+  //   }
+  // }
 
   const handleDelete = async (id) => {
-    deleteUser(id).then((data) => {
-      setUsers(() => users.filter((user) => user._id !== id))
-    })
+    const confirmed = window.confirm('Are you sure you want to delete this user?')
+    if (confirmed) {
+      deleteUser(id).then((data) => {
+        setUsers(() => users.filter((user) => user._id !== id))
+        alert('User deleted successfully.')
+      })
+    }
   }
+  // render
+  const displayedUsers = filteredUsers
+    .slice((pageIndex - 1) * pageSize, pageIndex * pageSize)
+    .map((user) => (
+      <tr key={user._id}>
+        <th scope="row">{user.email}</th>
+        <td>{user.username} </td>
+        <td>{user.isAdmin ? 'admin' : 'user'}</td>
+        <td>
+          {auth.isAdmin && (
+            <button className="btn btn-danger" onClick={() => handleDelete(user._id)}>
+              Delete
+            </button>
+          )}
+        </td>
+      </tr>
+    ))
   useEffect(() => {
     getAll().then((data) => setUsers(data.filter((x) => x._id !== auth._id)))
   }, [auth._id])
@@ -36,12 +68,7 @@ const User = () => {
       <div className="card-body">
         <CInputGroup className="mb-3">
           <CInputGroupText id="basic-addon1">search</CInputGroupText>
-          <CFormInput
-            onChange={(e) => setSearch(e.target.value)}
-            placeholder="Username"
-            aria-label="Username"
-            aria-describedby="basic-addon1"
-          />
+          <CFormInput onChange={(e) => setSearch(e.target.value)} placeholder="Username" />
         </CInputGroup>
         <div className="example">
           <div className="tab-content rounded-bottom">
@@ -49,37 +76,13 @@ const User = () => {
               <table className="table">
                 <thead>
                   <tr>
-                    <th scope="col">id</th>
                     <th scope="col">email</th>
                     <th scope="col">userName</th>
                     <th scope="col">IsAdmin</th>
                     <th scope="col">Acctions</th>
                   </tr>
                 </thead>
-                <tbody>
-                  {users
-                    .slice((pageIndex - 1) * pageSize, pageIndex * pageSize)
-                    .filter((user) => user.username.includes(search))
-                    .map((user) => (
-                      <tr key={user._id}>
-                        <th scope="row">{user._id}</th>
-
-                        <th scope="row">{user.email}</th>
-                        <td>{user.username} </td>
-                        <td>{user.isAdmin ? 'admin' : 'user'}</td>
-                        <td>
-                          {auth.isAdmin && (
-                            <button
-                              className="btn  btn-danger"
-                              onClick={() => handleDelete(user._id)}
-                            >
-                              Delete
-                            </button>
-                          )}
-                        </td>
-                      </tr>
-                    ))}
-                </tbody>
+                <tbody>{displayedUsers}</tbody>
               </table>
               <nav aria-label="Page navigation example">
                 <ul className="pagination">
